% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_dna.R
\name{boot_dna}
\alias{boot_dna}
\title{Boostrapping DNA sequences}
\usage{
boot_dna(dna, block = 1)
}
\arguments{
\item{dna}{an object of class \link[ape:DNAbin]{DNAbin}.}

\item{block}{integer. Number of columns to be resampled together. Default to 1.}
}
\value{
a \link[ape:DNAbin]{DNAbin} object.
}
\description{
\code{boot_dna()} generates random bootstrap alignments for confidence interval estimation
using \link{confidence_intervals}. Thus, it is meant to be an internal function of this package.
}
\examples{
boot <- boot_dna(geophagus)

}
\author{
Pedro S. Bittencourt
}
\keyword{internal}
