% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delim_tbl.R
\name{delim_tbl}
\alias{delim_tbl}
\title{Generating a Delimitation Hypothesis Table}
\usage{
delim_tbl(labels, sppVector, delimname = NULL)
}
\arguments{
\item{labels}{Vector of unique sequence ID labels.}

\item{sppVector}{Vector of corresponding morphological species delimitation groups.}

\item{delimname}{Character. User needs to supply string to name the delimitation method in the table. Default to NA.}
}
\value{
an object of class \link[tibble:tbl_df]{tbl_df}.
}
\description{
\code{delim_tbl()} returns species partition hypothesis estimated from a prior delimitation hypothesis supplied by the user.
}
\details{
\code{delim_tbl()} uses information in a species name vector to label each unique sample with a number corresponding to this name.
}
\examples{

# create a tibble
delim_df <- delim_tbl(
  labels = geophagus_info$gbAccession,
  sppVector = geophagus_info$scientificName,
  delimname = "SNAPPER"
)

# check
delim_df

}
\author{
Rupert A. Collins
}
