% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delimtools-package.R
\docType{package}
\name{delimtools-package}
\alias{delimtools}
\alias{delimtools-package}
\title{Helper Functions for Species Delimitation Analysis}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Helpers functions to process, analyse, and visualize the output of single locus species delimitation methods. For full functionality, please install suggested software at \url{https://legallab.github.io/delimtools/articles/install.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/legalLab/delimtools}
  \item \url{https://legallab.github.io/delimtools/}
  \item Report bugs at \url{https://github.com/legalLab/delimtools/issues}
}

}
\author{
\strong{Maintainer}: Pedro Bittencourt \email{pedro.sennabittencourt@gmail.com} (\href{https://orcid.org/0000-0002-4074-3541}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Pedro Bittencourt \email{pedro.sennabittencourt@gmail.com} (\href{https://orcid.org/0000-0002-4074-3541}{ORCID}) [copyright holder]
  \item Rupert Collins \email{rupertcollins@gmail.com} (\href{https://orcid.org/0000-0002-9135-1169}{ORCID}) [contributor, copyright holder]
  \item Tomas Hrbek \email{hrbek@evoamazon.net} (\href{https://orcid.org/0000-0003-3239-7068}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Tomas Flouri (mPTP C engine (AGPL-3)) [contributor]
  \item Sarah Lutteropp (mPTP C engine (AGPL-3)) [contributor]
  \item Paschalia Kapli (mPTP C engine (AGPL-3)) [contributor]
}

}
