% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgmyc.R
\name{.bgmyc_point}
\alias{.bgmyc_point}
\title{Internal: point-estimate partition from co-occurrence matrix}
\usage{
.bgmyc_point(probmat, tip_labels, ppcutoff, colname = "bgmyc")
}
\description{
Internal: point-estimate partition from co-occurrence matrix
}
\keyword{internal}
