% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgmyc.R
\name{.bgmyc_probmat}
\alias{.bgmyc_probmat}
\title{Internal: build co-occurrence probability matrix from retained samples}
\usage{
.bgmyc_probmat(par, setup, n_tips)
}
\description{
Internal: build co-occurrence probability matrix from retained samples
}
\keyword{internal}
