% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_sequences.R
\name{drop_sequences}
\alias{drop_sequences}
\title{Remove Sequences of a FASTA File}
\usage{
drop_sequences(fasta, drop)
}
\arguments{
\item{fasta}{a \code{\link[ape]{DNAbin}} object.}

\item{drop}{a character vector specifying the sequences to remove.}
}
\value{
an object of class \code{\link[ape]{"DNAbin"}}.
}
\description{
\code{drop_sequences()} removes sequences of a FASTA file by its names.
}
\details{
\code{drop_sequences()} relies on exact match between sequence names within
a fasta file and \code{drop} argument.
}
\author{
Pedro S. Bittencourt
}
