% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geophagus.R
\docType{data}
\name{geophagus}
\alias{geophagus}
\title{Cytochrome C Oxidase Sequences of Geophagus Eartheaters}
\format{
An object of class \link[ape:DNAbin]{DNAbin}
}
\source{
Ximenes AM, Bittencourt PS, Machado VN, Hrbek T, Farias IP. 2021.
Mapping the hidden diversity of the Geophagus sensu stricto species group
(Cichlidae: Geophagini) from the Amazon basin. PeerJ 9:e12443.
}
\usage{
geophagus
}
\description{
This is a set of 354 sequences of the mitochondrial gene cytochrome c oxidase
subunit I (COI-5P) of the eartheaters of the \emph{Geophagus sensu stricto} species group
downloaded from GenBank. Most of these sequences are from the data analysed by
Ximenes et al. (2021).
}
\keyword{datasets}
