% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geophagus.R
\docType{data}
\name{geophagus_beast}
\alias{geophagus_beast}
\title{Geophagus Eartheaters Ultrametric Tree}
\format{
An object of class \link[tidytree:treedata-class]{treedata}.
}
\usage{
geophagus_beast
}
\description{
This is a Maximum Clade Credibility (MCC) tree containing unique haplotypes from
\link{geophagus} estimated using BEAST2 v2.6.7. Unique haplotypes
were select using \link{hap_collapse}.
}
\keyword{datasets}
