% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geophagus.R
\docType{data}
\name{geophagus_bootstraps}
\alias{geophagus_bootstraps}
\title{Geophagus Eartheaters Bootstrap Trees}
\format{
An object of class \link[ape:multiphylo]{multiphylo}
}
\usage{
geophagus_bootstraps
}
\description{
This is a set of 100 Maximum Likelihood trees sampled from bootstrap trees used to
estimate \link{geophagus_raxml} using RAxML-NG v. 1.1.0-master.
Meant to be used for \link{confidence_intervals} estimation.
}
\keyword{datasets}
