% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geophagus.R
\docType{data}
\name{geophagus_delims}
\alias{geophagus_delims}
\title{Geophagus Eartheaters Species Partitions}
\format{
A dataframe with 137 rows and 9 columns:
\describe{
\item{labels}{Unique haplotype labels}
\item{abgd}{species partitions for \code{ABGD} method}
\item{asap}{species partitions for \code{ASAP} method}
\item{bgmyc}{species partitions for \code{bGMYC} method}
\item{gmyc}{species partitions for \code{GMYC} method}
\item{locmin}{species partitions for \code{locmin} method}
\item{morph}{species partitions following NCBI taxonomy}
\item{mptp}{species partitions for \code{mPTP} method}
\item{ptp}{species partitions for \code{PTP} method}
}
}
\usage{
geophagus_delims
}
\description{
This is a data frame containing species delimitation partitions for all the
137 unique haplotypes of \link{geophagus} generated using
functions contained in this package. Use \link{report_delim}
to check number of lineages per method.
}
\keyword{datasets}
