% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geophagus.R
\docType{data}
\name{geophagus_info}
\alias{geophagus_info}
\title{Geophagus Earthearts Associated Metadata}
\format{
A data frame with 354 rows and 19 columns:
\describe{
\item{scientificName}{scientific name}
\item{scientificNameGenBank}{scientific name following NCBI taxonomy}
\item{class}{class}
\item{order}{order}
\item{family}{family}
\item{genus}{genus}
\item{dbid}{NCBI Nucleotide Database internal ID}
\item{gbAccession}{NCBI Nucleotide Database accession number}
\item{gene}{Gene acronym}
\item{length}{Sequence length in base pairs (bp)}
\item{organelle}{Organelle from which gene was sequenced}
\item{catalogNumber}{An identifier for the record within a data set or collection}
\item{country}{Name of the Country followed by sampling locality (when available)}
\item{publishedAs}{Title of the article which generated the sequences}
\item{publishedIn}{Journal which published the article}
\item{publishedBy}{A person, group, or organization responsible for depositing the sequence}
\item{date}{Date published}
\item{decimalLatitude}{Latitude in decimal degrees}
\item{decimalLongitude}{Longitude in decimal degrees}
}
}
\usage{
geophagus_info
}
\description{
This is the associated metadata for the 354 sequences of the mitochondrial gene
cytochrome c oxidase subunit I (COI-5P) of the \emph{Geophagus sensu stricto} species
group downloaded from GenBank and stored in \link{geophagus}.
}
\keyword{datasets}
