% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geophagus.R
\docType{data}
\name{geophagus_posterior}
\alias{geophagus_posterior}
\title{Geophagus Eartheaters Posterior Trees}
\format{
An object of class \link[ape:multiphylo]{multiphylo}
}
\usage{
geophagus_posterior
}
\description{
This is a set of 100 ultrametric trees sampled from the posterior trees used to
estimate \link{geophagus_beast} using BEAST2 v2.6.7. Meant to be
used for \link{confidence_intervals} estimation.
}
\keyword{datasets}
