% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geophagus.R
\docType{data}
\name{geophagus_raxml}
\alias{geophagus_raxml}
\title{Geophagus Eartheaters Phylogram}
\format{
An object of class \link[tidytree:treedata-class]{treedata}.
}
\usage{
geophagus_raxml
}
\description{
This is a Maximum Likelihood Estimation Tree containing unique haplotypes from
\link{geophagus} estimated using RAxML-NG v. 1.1.0-master.
Unique haplotypes were select using \link{hap_collapse}.
}
\keyword{datasets}
