% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mptp.R
\name{mptp_support}
\alias{mptp_support}
\title{Extract MCMC support values}
\usage{
mptp_support(x)
}
\arguments{
\item{x}{An object of class \code{"mptp_mcmc"}.}
}
\value{
A data frame with columns \code{node_label} and \code{support}.
}
\description{
Extract MCMC support values
}
