% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgd_tbl.R
\name{abgd_tbl}
\alias{abgd_tbl}
\title{A Command-Line Interface for ABGD - Automatic Barcode Gap Discovery}
\usage{
abgd_tbl(
  infile,
  exe = NULL,
  haps = NULL,
  slope = 1.5,
  model = 3,
  outfolder = NULL
)
}
\arguments{
\item{infile}{Path to fasta file.}

\item{exe}{Path to an ABGD executable.}

\item{haps}{Optional. A vector of haplotypes to keep into the \code{\link[tibble]{tbl_df}}.}

\item{model}{An integer specifying evolutionary model to be used. Available options are:
\itemize{
  \item 0: Kimura-2P
  \item 1: Jukes-Cantor (default)
  \item 2: Tamura-Nei
  \item 3: simple distance (p-distance)
}}

\item{outfolder}{Path to output folder. Default to NULL. If not specified, a temporary location is used.}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}
}
\description{
\code{abgd_tbl()} returns species partition hypothesis estimated by ABGD software
(\url{https://bioinfo.mnhn.fr/abi/public/abgd/}).
}
\details{
\code{abgd_tbl()} relies on \code{\link[base]{system}} to invoke ABGD software through
a command-line interface. Hence, you must have the software available as an executable file on 
your system in order to use this function properly.
\code{abgd_tbl()} saves all output files in \code{outfolder} and imports the first recursive partition
file generated to \code{Environment}.
}
\examples{
abgd_df <- abgd_tbl(infile = "data/pacus/pacus.trimmed.fasta", model= 3, outfolder = "pacus.asap")

}
\author{
N. Puillandre,  A. Lambert,  S. Brouillet,  G. Achaz
}
