% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dwc.R
\name{as_dwc}
\alias{as_dwc}
\title{Rename Columns using Darwin Core Standard Terms}
\usage{
as_dwc(dwc, data, terms)
}
\arguments{
\item{dwc}{a list of standard terms and definitions created using \code{\link[delimtools]{get_dwc}}}

\item{data}{a \code{\link[tibble]{tbl_df}}}

\item{terms}{a vector or list of terms to be used as replacement.}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}.
}
\description{
\code{as_dwc()} rename columns in a \code{\link[tibble]{tbl_df}} using a vector of terms defined by
Darwin Core Standard.
}
\details{
\code{as_dwc()} will replace current column names by the ones defined in \code{terms}. For each
column in \code{data}, Darwin Core equivalent terms must be informed in the same order
by the user. If \code{terms} and column names do not match in length or if \code{terms} used
are not found in Darwin Core standard, an error will be printed on Console.
}
\author{
Pedro S. Bittencourt, Rupert A. Collins.
}
