% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asap_tbl.R
\name{asap_tbl}
\alias{asap_tbl}
\title{A Command-Line Interface for ASAP - Assemble Species by Automatic Partitioning}
\usage{
asap_tbl(infile, exe = NULL, haps = NULL, model = 3, outfolder = NULL)
}
\arguments{
\item{infile}{Path to fasta file.}

\item{exe}{Path to an ASAP executable.}

\item{haps}{Optional. A vector of haplotypes to keep into the \code{\link[tibble]{tbl_df}}.}

\item{model}{An integer specifying evolutionary model to be used. Available options are:
\itemize{
  \item 0: Kimura-2P
  \item 1: Jukes-Cantor (default)
  \item 2: Tamura-Nei
  \item 3: simple distance (p-distance)
}}

\item{outfolder}{Path to output folder. Default to NULL. If not specified, a temporary location is used.}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}
}
\description{
\code{asap_tbl()} returns species partition hypothesis estimated by ASAP software
(\url{https://bioinfo.mnhn.fr/abi/public/asap/}).
}
\details{
\code{asap_tbl()} relies on \code{\link[base]{system}} to invoke ASAP software through
a command-line interface. Hence, you must have the software available as an executable file on 
your system in order to use this function properly.
\code{asap_tbl()} saves all output files in \code{outfolder} and imports the first partition
file generated to \code{Environment}.
}
\examples{
asap_df <- asap_tbl(infile = "data/pacus/pacus.trimmed.fasta", model= 3, outfolder = "pacus.asap")

}
\author{
Nicolas Puillandre, Sophie Brouillet, Guillaume Achaz.
}
