% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgmyc_tbl.R
\name{bgmyc_tbl}
\alias{bgmyc_tbl}
\title{Turns bGMYC Results Into a Tibble}
\usage{
bgmyc_tbl(bgmyc_res, ppcutoff)
}
\arguments{
\item{bgmyc_res}{Output from \code{\link[bGMYC]{bgmyc.singlephy}}.}

\item{ppcutoff}{Posterior probability threshold for clustering samples into
species partitions. See \code{\link[bGMYC]{bgmyc.point}} for details.}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}.
}
\description{
\code{bgmyc_tbl()} processes output from \code{\link[bGMYC]{bgmyc.singlephy}} into an
object of class \code{\link[tibble]{tbl_df}}.
}
\details{
\code{\link[bGMYC]{bGMYC}} package uses \code{\link[bGMYC]{spec.probmat}} to create a
matrix of probability of conspecificity and \code{\link[bGMYC]{bgmyc.point}}
to split samples into a list which individuals
meets the threshold specified by \code{ppcutoff}. \code{bgmyc_tbl()} wraps up these
two functions into a single one and turns these inputs into a tibble which matches
the output from \code{\link[delimtools]{gmyc_tbl()}} and \code{\link[delimtools]{locmin_tbl()}}
}
\author{
Noah M. Reid.
}
