% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_others.R
\name{collapse_others}
\alias{collapse_others}
\title{Summarise Haplotype Metadata Down to One Row}
\usage{
collapse_others(data, hap_tbl, identifier, cols)
}
\arguments{
\item{data}{an object of class \code{\link[tibble]{tbl_df}} containing sequence metadata.}

\item{hap_tbl}{output from \code{link[delimtools]{haplotype_tbl()}}.}

\item{identifier}{a character vector of unique sequence labels.}

\item{cols}{a character vector of variables to collapse.}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}.
}
\description{
\code{collapse_others()} returns a \code{\link[tibble]{tbl_df}} summarising
all unique haplotype frequencies, duplicates and selected metadata into a single row.
}
\details{
\code{collapse_others()} is a helper function to summarise metadata along with
\code{\link[delimtools]{haplotype_tbl()}}. For any given \code{cols},
\code{collapse_others()} flattens its content by unique haplotypes and its duplicates
in \code{hap_tbl}.
}
\author{
Pedro S. Bittencourt, Rupert A. Collins.
}
