% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delim_brewer.R
\name{delim_brewer}
\alias{delim_brewer}
\title{Customize Delimitation Colors}
\usage{
delim_brewer(delim, name, n, seed)
}
\arguments{
\item{delim}{Output from \code{\link[delimtools]{delim_join}}.}

\item{name}{A palette name. See \code{\link[RColorBrewer]{RColorBrewer}} for options.}

\item{n}{Number of different colors to interpolate. Minimum 3 and maximum depending on palette.}
}
\description{
\code{delim_brewer()} returns a set of colors created by interpolating
color palettes from \code{\link[RColorBrewer]{RColorBrewer}}.
}
\details{
\code{delim_brewer()} interpolates over a color palette and returns a vector of random colors
whose length is equal to the sum of unique species delimitation partitions in \code{delim}.
For reproducibility, make sure to use in combination with \code{\link[base]{set.seed}}.
One should also try different seeds to get best color combinations for plotting.
}
\author{
Rupert A. Collins, Pedro S. Bittencourt
}
