% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delim_join.R
\name{delim_join}
\alias{delim_join}
\title{Join Multiple Species Delimitation Methods Outputs}
\usage{
delim_join(delim)
}
\arguments{
\item{delim}{A \code{\link[base]{list}} or \code{\link[base]{data.frame}} of multiple
species delimitation methods outputs.}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}.
}
\description{
\code{delim_join()} returns a \code{\link[tibble]{tbl_df}} of species delimitation
outputs whose partitions are consistent across different methods.
}
\details{
\code{delim_join()} is a helper function to join multiple lists or columns of species
delimitation outputs into a single \code{\link[tibble]{tbl_df}} while keeping consistent
identifications across multiple methods. Species delimitation outputs are in general a
list or data frame of sample labels and its species partitions (Species 1, Species 2, etc.). These
partition names may be or not the same across two or more methods. \code{delim_join()} standardizes
partition names across two or more species delimitation outputs while keeping its underlying structure intact.
}
\author{
Pedro S. Bittencourt, Rupert A. Collins.
}
