% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delimtools-package.R
\docType{package}
\name{delimtools-package}
\alias{delimtools}
\alias{delimtools-package}
\title{Helper Functions for Species Delimitation Analysis}
\description{
This package contains helpers functions to process, analyse, and plot the output of single locus species delimitation methods. The main purpose of delimtools is to facilitate some work routines for the average R user, from the organization of data to plotting trees and associated data as easily as possible.
}
\author{
\strong{Maintainer}: Pedro Bittencourt \email{pedro.sennabittencourt@gmail.com} (\href{https://orcid.org/0000-0002-4074-3541}{ORCID})

Authors:
\itemize{
  \item Rupert Collins \email{rupertcollins@gmail.com} (\href{https://orcid.org/0000-0002-9135-1169}{ORCID}) [contributor]
  \item Tomas Hrbek \email{hrbek@evoamazon.net} (\href{https://orcid.org/0000-0003-3239-7068}{ORCID}) [contributor]
}

}
