% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_sequences.R
\name{drop_sequences}
\alias{drop_sequences}
\title{Remove Sequences of a DNAbin list object}
\usage{
drop_sequences(fasta, identifier, drop = TRUE)
}
\arguments{
\item{fasta}{a \code{\link[ape]{DNAbin}} object.}

\item{identifier}{a character vector containing sequence names.}

\item{drop}{Logical. If \code{TRUE}, sequence names in \code{identifier} will
be dropped from \code{fasta}. If \code{FALSE}, sequence names absent in \code{identifier}
will be dropped instead.}
}
\value{
an object of class \code{\link[ape]{"DNAbin"}}.
}
\description{
\code{drop_sequences()} removes sequences of a FASTA file by its names.
}
\details{
\code{drop_sequences()} relies on exact match between sequence names within
a fasta file and \code{identifier} argument.
}
\examples{

# Import data from ape package.
data(woodmouse, package = "ape")

# Create a vector of sequence names to drop or keep.
identifier <- dimnames(woodmouse)[[1]][1:3]

# Remove sequences listed in identifier
drop_sequences(woodmouse, identifier, drop = TRUE)

# Remove sequences not listed in identifier
drop_sequences(woodmouse, identifier, drop = FALSE)

}
\author{
Pedro S. Bittencourt
}
