% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmyc_tbl.R
\name{gmyc_tbl}
\alias{gmyc_tbl}
\title{Turns GMYC Results Into a Tibble}
\usage{
gmyc_tbl(gmyc_res)
}
\arguments{
\item{gmyc_res}{Output from \code{\link[splits]{gmyc}}.}
}
\value{
An object of class \code{\link[tibble]{tbl_df}}.
}
\description{
\code{gmyc_tbl()} processes output from \code{\link[splits]{gmyc}} into an
object of class \code{\link[tibble]{tbl_df}}.
}
\details{
\code{\link[splits]{splits}} package uses \code{\link[splits]{gmyc}} to optimize
genetic clusters and \code{\link[splits]{spec.list}} to cluster samples into
species partitions. \code{gmyc_tbl()} turns these results into a tibble which matches
the output from \code{\link[delimtools]{bgmyc_tbl()}} and \code{\link[delimtools]{locmin_tbl()}}.
}
\author{
Thomas Ezard, Tomochika Fujisawa, Tim Barraclough.
}
