% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_tbl.R
\name{haplotype_tbl}
\alias{haplotype_tbl}
\title{Summarise Haplotypes Down to One Row}
\usage{
haplotype_tbl(dna, verbose = TRUE)
}
\arguments{
\item{dna}{an object of class \code{\link[ape]{DNAbin}}.}

\item{verbose}{logical. Returns a warning if any sequence contains non ACTG bases.
See \code{link[delimtools]{clean_dna}} for details.}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}.
}
\description{
\code{haplotype_tbl()} returns a \code{\link[tibble]{tbl_df}} summarising
all unique haplotype frequencies and duplicates into a single row.
}
\details{
\code{haplotype_tbl()} uses a combination of \code{\link[delimtools]{clean_dna}} and
\code{\link[delimtools]{hap_collapse}} to summarise haplotypes into a tibble. Each row
of the tibble has an unique haplotype, its frequency and all its collapsed duplicates in a
flattened string.
}
\author{
Rupert A. Collins, Pedro S. Bittencourt.
}
