% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locmin_tbl.R
\name{locmin_tbl}
\alias{locmin_tbl}
\title{Turns Local Minima Results into a Tibble}
\usage{
locmin_tbl(distobj, threshold = 0.01, haps = NULL, delimname = "locmin")
}
\arguments{
\item{distobj}{A distance object (usually from \code{\link[ape]{dist.dna}}).}

\item{threshold}{Distance cutoff for clustering. Default of 0.01. See
\code{\link[spider]{localMinima}} for details.}

\item{haps}{Optional. A vector of haplotypes to keep into the \code{\link[tibble]{tbl_df}}.}

\item{delimname}{Character. String to rename the delimitation method in the table. Default to 'locmin'.}
}
\value{
An object of class \code{\link[tibble]{tbl_df}}.
}
\description{
\code{locmin_tbl()} processes output from \code{\link[spider]{tclust}} into
an object of class \code{\link[tibble]{tbl_df}}.
}
\details{
\code{\link[spider]{spider}} package uses \code{\link[spider]{localMinima}} to
determine possible thresholds for any distance matrix and \code{\link[spider]{tclust}}
to cluster samples within a given \code{threshold} into species partitions.
\code{locmin_tbl()} turns these inputs into a tibble which matches
the output from \code{\link[delimtools]{gmyc_tbl()}} and \code{\link[delimtools]{bgmyc_tbl()}}.
}
\author{
Samuel Brown.
}
