% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_brlen.R
\name{min_brlen}
\alias{min_brlen}
\title{A function to report the smallest tip-to-tip distances in a phylogenetic tree}
\usage{
min_brlen(tree, n = 5, print = TRUE)
}
\arguments{
\item{tree}{A path to tree file in Newick format, or a phylogenetic tree object of class "phylo".}

\item{n}{Number of distances to report (default = 5).}

\item{print}{Logical of whether to print the result to screen (default = TRUE).}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}
}
\description{
\code{min_brlen()} returns a table of smallest tip-to-tip distances in a phylogenetic tree.
}
\details{
\code{min_brlen()} tabulates the smallest tip-to-tip distances in a phylogenetic tree
using \code{ape::cophenetic.phylo()} and prints a table to screen.
This is useful when excluding identical or near-identical haplotypes
using the '--minbr' parameter in mPTP.
}
\author{
Rupert A. Collins
}
