% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_tbl.R
\name{morph_tbl}
\alias{morph_tbl}
\title{Generating a morphological delimitation table}
\usage{
morph_tbl(labels, sppVector, delimname = "morph")
}
\arguments{
\item{labels}{Vector of unique sequence ID labels.}

\item{sppVector}{Vector of corresponding morphological species delimitation groups.}

\item{delimname}{Character. String to rename the delimitation method in the table. Default to 'morph'.}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}
}
\description{
\code{morph_tbl()} returns species partition hypothesis estimated from a prior taxonomic identifications supplied by the user.
}
\details{
\code{morph_tbl()} uses information in a species name vector to label each unique sample with a number corresponding to this name.
}
\examples{
morph.df <- morph_tbl(labels = c("seq1", "seq2", "seq3", "seq4"), 
                      sppVector = c("spA", "spB", "spC", "spC"))

}
\author{
Rupert A. Collins
}
