% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mptp.R
\name{mptp}
\alias{mptp}
\title{A Command-Line Interface for mPTP - multi-rate Poisson Tree Processes}
\usage{
mptp(infile, outfolder, method = c("multi", "single"), outgroup = NULL)
}
\arguments{
\item{infile}{Path to tree file in Newick format.}

\item{outfolder}{Path to output folder.}

\item{method}{Which algorithm for Maximum Likelihood point-estimate to be used. Available options are:
\itemize{
\item single Single-rate PTP model. It assumes that every species evolved with the same rate.
\item multi Multi-rate mPTP model. It assumes that all species have different evolutionary rates.
}}

\item{outgroup}{A vector of outgroup labels. Default to NULL. If specified,
mPTP will crop the specified labels from the tree.}
}
\value{
an object of class \code{\link[tibble]{tbl_df}}
}
\description{
\code{mptp()} returns species partition hypothesis estimated by mPTP software
(\url{https://github.com/Pas-Kapli/mptp}).
}
\details{
\code{mptp()} relies on \code{\link[base]{system}} to invoke mPTP software through
a command-line interface. Hence, you must have the software installed on your
operating system root in order to use this function properly. \code{mptp()}
saves all output files in \code{outfolder} and imports the results generated to \code{Environment}.
}
\author{
Paschalia Kapli, Sarah Lutteropp, Jiajie Zhang, Kassian Kobert, Pavlos Pavlides, Alexandros Stamatakis, Tomáš Flouri.
}
